### 2019-07-18 #1583 @whedon
[PRE REVIEW]: ivis: dimensionality reduction in very large datasets using Siamese Networks
**Submitting author:** @beringresearch (<a href="http://orcid.org/0000-0001-6727-4688">Ignat Drozdov</a>)
**Repository:** <a href="https://github.com/beringresearch/ivis" target ="_blank">https://github.com/beringresearch/ivis</a>
**Version:** v1.2.3
