### 2019-07-23 #1583 @lpantano
Thank you, @SaskiaFreytag, I will add you as a reviewer, then I need to find another one (if you know somebody would like to do it, let me know). At that moment, I will start another issue where the review happens (mainly a list of items to check, like https://github.com/openjournals/joss-reviews/issues/1584.
