### 2019-07-23 4f8755b
Enable running ivis on Windows. Disable building on Annoy Index on disk when on Windows platform.
