### 2019-07-23 8311382
Update AnnoyIndex object to supply metric parameter angular to stay in line with future AnnoyIndex API updates.
