### 2019-07-28 beringresearch/ivis#28 @idroz
Hi - thanks for trying out `ivis` and testing it out in RStudio. It does sound like a path issue. I haven't worked much with RStudio recently, but it may be that RStudio doesn't inherit system $PATH: https://stackoverflow.com/questions/31121645/rstudio-shows-a-different-path-variable

I'm thinking that a potential solution would be to get rid of Conda manager altogether and simply install R wrapper through virtualenv + pip. I'll test this out over the next few days and revert here.

Thanks again!
