### 2019-07-28 f027cb5
migraite conda to virtualenv
