### 2019-07-28 7417b71
migraite conda to virtualenv
