### 2019-07-28 8c7cb2d
catch up R with python changes
