### 2019-07-28 beringresearch/ivis#28 @kevinrue
To be fair, I am not an expert with regard to `reticulate` yet, so there may be some trick to fix the situation that I don't know. I'd recommend looking at other packages that use `reticulate` to see how they deal with this kind of situation, if you can find some.
I look forward to testing the method directly within an R environment, but in the meantime I'm happy to test the functionality in a Python environment.
Good luck!
