### 2019-07-28 beringresearch/ivis#28 @idroz
Hi again - I made a few changes to `ivis` install scripts. They now use `virtualenv` rather than Conda. I've also caught up R wrapper with the latest stable version of `ivis` on PyPi. Both installation and demo scripts seem to work inside RStudio and on command line. Could you please test out the development branch:

```
devtools::install_github("beringresearch/ivis/R-package", ref="r-virtualenv-install")
```

Subsequent are the same as in master branch README.
Thanks!
