### 2019-07-28 #1596 @kevinrue
# General checks
For the record, the submitting author (@beringresearch) is an institutional account representative of the affiliation of the authors who made major contributions to the software.
This is frequent practice and absolutely fine.
