### 2019-07-28 beringresearch/ivis#28 @kevinrue
After updating my installation using your `devtools`  command above, the following worked for me now, well done!
```
library(ivis)

model <- ivis(k = 3, batch_size = 3)

X = data.matrix(iris[, 1:4])
model = model$fit(X)

xy = model$transform(X)
```
