### 2019-07-28 beringresearch/ivis#29 @kevinrue
Solved simultaneously to #28
