### 2019-07-28 beringresearch/ivis#29 @kevinrue
I'll reopen and let you close the issue when you merge the update in a PR.
