### 2019-07-28 beringresearch/ivis#28 @idroz
Addressed in #30
