### 2019-07-29 beringresearch/ivis#32 @SaskiaFreytag
Attempt to apply to non-function
Hi,

I am reviewing for [JOSS-Reviews](https://github.com/openjournals/joss-reviews/issues/1596).

I think I was able to install the package. Unfortunately when I run the example code, I run into the following error:

> model <- ivis(k = 3)
Error in ivis_object$Ivis(embedding_dims = embedding_dims, k = k, distance = distance,  :
  attempt to apply non-function`

My session info:
```
> sessionInfo()
R version 3.6.0 (2019-04-26)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Mojave 10.14.4

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRlapack.dylib

locale:
[1] en_AU.UTF-8/en_AU.UTF-8/en_AU.UTF-8/C/en_AU.UTF-8/en_AU.UTF-8

attached base packages:

other attached packages:
[1] ivis_1.2.3

[1] compiler_3.6.0  Matrix_1.2-17   tools_3.6.0     yaml_2.2.0      Rcpp_1.0.2      reticulate_1.13
[7] grid_3.6.0      jsonlite_1.6    lattice_0.20-38
```
