### 2019-07-29 beringresearch/ivis#32 @idroz
Hi - thanks for trying out ivis!
It looks like the install may not have gone correctly on the Python side. Since R is just a wrapper around the python package, it may be unable to import the ivis library.

Do you get any errors when installing ivis from PyPi?

```
pip install ivis
```

 and then

```python
from ivis import Ivis
```

to make sure everything worked.
