### 2019-07-29 beringresearch/ivis#32 @SaskiaFreytag
It works now. I think maybe adding this to the installation guide would be helpful.

Also one way to set the environment variable in R is:
`Sys.setenv(KERAS_BACKEND = "tensorflow")`
