### 2019-07-29 beringresearch/ivis#33 @SaskiaFreytag
Consider adding automated tests
Also part of the [JOSS-Review](https://github.com/openjournals/joss-reviews/issues/1596).

Please consider adding [automated tests](http://r-pkgs.had.co.nz/tests.html) for the R package.
