### 2019-07-29 beringresearch/ivis#34 @SaskiaFreytag
Add contributor guidelines
This is part of the [JOSS-Review](https://github.com/openjournals/joss-reviews/issues/1596)

Please consider adding clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
