### 2019-07-29 beringresearch/ivis#35 @SaskiaFreytag
Extend readme in R package
This is part of [JOSS-Review](https://github.com/openjournals/joss-reviews/issues/1596)

It would be nice to showcase the computed ivis visualisation of the data in R package. Consider adding the following code:

```
library(ggplot2)

dat <- data.frame(x=xy[,1],
                  y=xy[,2],
                  species=iris$Species)

ggplot(dat, aes(x=x, y=y)) + geom_point(aes(color=species)) + theme_classic()
```
