+++ .github/ISSUE_TEMPLATE/support.md
@@ -1,0 +1,10 @@
+---
+name: Support
+about: General questions about `ivis`
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+General questions about algorithm design and usage.
