### 2019-07-29 f59e835
Create CONTRIBUTING.md
contributing guidelines
