### 2019-07-29 beringresearch/ivis#34 @idroz
Hi - I added contributing guidelines and issue report templates to the master branch. Let me know if OK!
