### 2019-07-29 9a215a2
add scatterplot example
