### 2019-07-29 2207517
update code snippet syntax
