### 2019-07-29 beringresearch/ivis#35 @idroz
Updated R README with the plotting snippet - thanks!
