### 2019-07-29 beringresearch/ivis#34 @kevinrue
Suggestion for further improving the [CONTRIBUTING.md](https://github.com/beringresearch/ivis/blob/master/CONTRIBUTING.md) file:

> If your PR introduces a change in functionality, make sure you start by writing a design doc.

I suggest providing a template, or at least a minimal description of what you expect as a "design doc".
Without providing a template, you create extra burden on the contributor (guessing) and yourself (explaining) or you could simply scare off prospective contributors.

> Make sure any code you touch still has up-to-date docstrings and documentation.

Suggestion: "any code you _edit_"

> Write tests.

I suggest inviting contributors to look at https://github.com/beringresearch/ivis/tree/master/tests to learn and follow your testing style.
