### 2019-07-29 beringresearch/ivis#33 @kevinrue
It may be slightly more difficult to set up than "pure" R packages, but once your first R tests pass successfully on your local machine, I also recommend using [Travis CI](https://travis-ci.org) to run those tests automatically and display a [badge](https://docs.travis-ci.com/user/status-images/) in the README file to track and advertise the result of those tests.
