### 2019-07-30 ad8e051
add links to template docs
