### 2019-07-30 beringresearch/ivis#34 @idroz
Thanks @kevinrue . I've added the suggested changes, including a link to the design doc template, to the [CONTRIBUTING.md](https://github.com/beringresearch/ivis/blob/master/CONTRIBUTING.md) file.
