### 2019-07-30 e168f9d
add R-package test
