### 2019-07-30 73c0a39
Add cosine distance metric for embedding
