### 2019-07-30 346b05d
Add cosine distance metric for embedding
