### 2019-07-30 beringresearch/ivis#38 @kevinrue
Add a vignette to the R package
Hello,

For [JOSS review](https://github.com/openjournals/joss-reviews/issues/1596).

The R package lacks documentation of an application to a real-life dataset.

**Describe the solution you'd like**

Please add a [vignette](http://r-pkgs.had.co.nz/vignettes.html) in the R package demonstrating at least an example application to a single-cell dataset. Basically, the equivalent of the scanpy workflow [here](https://bering-ivis.readthedocs.io/en/latest/scanpy_singlecell.html).

A convenient way to use the `pbmc3k` dataset for demonstration purposes is the Bioconductor [TENxPBMCData](https://bioconductor.org/packages/release/data/experiment/html/TENxPBMCData.html) package.

Suggested code:
```r
library(TENxPBMCData)
tenx_ pbmc3k <- TENxPBMCData(dataset = "pbmc3k")
```

Ideally, consider using the vignette (or a separate one) to also give an introduction to the functionality of the R package.
It is not necessary to duplicate information already described in the documentation of the Python package (DRY principle); you may simply include a link to the [main page](https://bering-ivis.readthedocs.io/en/latest/index.html).

A working example of an R workflow could also be included in the documentation of the Python package, although this is probably unnecessarily difficult to maintain.
Ideally, that example would be run and tested for every new release of the Python and R source code.

**Additional context**
Once you have an R vignette written, you should also consider using [pkgdown](https://pkgdown.r-lib.org) to automatically create a GitHub website including the full package documentation.
