### 2019-07-30 eebe7e9
add r vignette
