+++ R-package/vignettes/ivis_singlecell.Rmd
@@ -25,7 +25,7 @@ cd data; tar -xzf pbmc3k_filtered_gene_bc_matrices.tar.gz
 ```{r setup}
 library(Seurat)
 
-pbmc.data <- Read10X(data.dir = "../data/filtered_gene_bc_matrices/hg19/")
+pbmc.data <- Read10X(data.dir = "data/filtered_gene_bc_matrices/hg19/")
 pbmc <- CreateSeuratObject(counts = pbmc.data, project = "pbmc3k",
                            min.cells = 3, min.features = 200)
 pbmc
