### 2019-07-30 beringresearch/ivis#38 @idroz
Hi there - I've added an [R vignette](https://github.com/beringresearch/ivis/tree/master/R-package/vignettes) that demonstrates how `ivis` can be integrated into a Seurat workflow. The vignette uses an example single cell dataset mentioned earlier.

With respect to pkgdown, as the main package implementation is in python, we feel that readthedocs is more appropriate for the project. As documentation expands, we will incorporate R vignette into the readthedocs builds.
