+++ R-package/NAMESPACE
@@ -4,5 +4,5 @@ export(install_ivis)
 export(ivis)
 importFrom(reticulate,py_install)
 importFrom(reticulate,virtualenv_create)
-importFrom(reticulate,virtualenv_install)
 importFrom(reticulate,virtualenv_list)
+importFrom(reticulate,virtualenv_remove)
