### 2019-07-30 beringresearch/ivis#38 @kevinrue
Thanks!

I could run the R vignette manually, but that required following it on your [repository](https://github.com/beringresearch/ivis/blob/master/R-package/vignettes/ivis_singlecell.Rmd#L19), and manually running the chunk of bash code to download the raw data from 10X (https://github.com/beringresearch/ivis/blob/master/R-package/vignettes/ivis_singlecell.Rmd#L19).

One of the main advantages of an R vignette is that it is generally compiled and installed with the R package itself (i.e., `browseVignettes("ivis")` should display the vignettes available).
This allows users to preview the package functionality before installing it, but also demonstrates that the package simply works.

To do so, as long as the R package is distributed directly from GitHub, I suggest that you update the installation command to the following, in order to build the vignette on the users' machine.
```r
devtools::install_github("beringresearch/ivis/R-package", build_vignettes = TRUE)
```

That said, the command above currently crashes because the vignette includes a chunk of code that needs to be run manually to  download the demonstration dataset (https://github.com/beringresearch/ivis/blob/master/R-package/vignettes/ivis_singlecell.Rmd#L19)

I will shortly open a PR to demonstrate the use of `TENxPBMCData` (mentioned above).

Having tested it locally, I must point out that installing the R package with `build_vignettes = TRUE` mentioned above will take longer to run, as it basically needs to run the method on the PBMC 3k dataset.
I haven't timed it precisely, but the overall process took a few minutes on my laptop.

I suggested `pkgdown` only because it automatically compiles and serves the R man pages and vignettes for free every time the R package is updated.
That said, you probably want to recompile the doc whenever any one of the Python or R package is updated, so I agree that readthedocs is more appropriate as the central documentation for this project.

If you can also compile and serve the R vignette there, I've got nothing to add.
That said, readthedocs is probably the best solution to update the documentation
