### 2019-07-30 d561de1
Update the vignette to fully compile automatically. Update list of suggested packages.
