+++ R-package/vignettes/ivis_singlecell.Rmd
@@ -14,7 +14,10 @@ knitr::opts_chunk$set(
 )
 ```
 
-This example will demonstrate how `ivis` can be used to visualise single cell experiments. Data import, preprocessing and normalisation are handled by the Seurat library. The data that will be used in this example consists of 3,000 PBMCs from a healthy donor and is freely available from 10x Genomics. Now, let’s download the data and get started.
+This example will demonstrate how `ivis` can be used to visualise single cell experiments.
+Data import, preprocessing and normalisation are handled by the Seurat library.
+The data that will be used in this example consists of 3,000 PBMCs from a healthy donor and is freely available from 10x Genomics.
+The data can be download manually in the following way.
 
 ```{bash, eval=FALSE}
 mkdir data
@@ -22,11 +25,20 @@ wget http://cf.10xgenomics.com/samples/cell-exp/1.1.0/pbmc3k/pbmc3k_filtered_gen
 cd data; tar -xzf pbmc3k_filtered_gene_bc_matrices.tar.gz
 ```
 
+Alternatively, the [`TENxPBMCData`](https://bioconductor.org/packages/release/data/experiment/html/TENxPBMCData.html) Bioconductor package provides tools to download and cache locally each of the 10x Genomics datasets.
+
+```{r, message=FALSE}
+library(TENxPBMCData)
+pbmc3k.sce <- TENxPBMCData(dataset = "pbmc3k")
+pbmc3k.sce
+```
+
+The data is obtained in the `SingleCellExperiment` class and can be converted to a `Seurat` object as follows.
+
 ```{r setup}
 library(Seurat)
-
-pbmc.data <- Read10X(data.dir = "data/filtered_gene_bc_matrices/hg19/")
-pbmc <- CreateSeuratObject(counts = pbmc.data, project = "pbmc3k",
+colnames(pbmc3k.sce) <- paste0("Cell", seq_len(ncol(pbmc3k.sce)))
+pbmc <- CreateSeuratObject(counts = assay(pbmc3k.sce, "counts"), project = "pbmc3k",
                            min.cells = 3, min.features = 200)
 pbmc
 ```
@@ -64,7 +76,7 @@ pbmc <- FindClusters(pbmc, resolution = 0.5)
 
 We are now ready to apply `ivis` to the PBMC dataset.
 
-```{r ivis, results = "hide"}
+```{r ivis, results = "hide", message=FALSE}
 library(ivis)
 
 model = ivis(embedding_dims=2, k=15, model='maaten',
@@ -81,7 +93,7 @@ pbmc[["ivis"]] <- CreateDimReducObject(embeddings = embeddings,
 ```
 
 ```{r plot, fig.height=6, fig.width=7}
-DimPlot(pbmc, reduction="ivis")
+DimPlot(pbmc, reduction="ivis", pt.size = 1)
 ```
 
 `ivis` effectively captured distinct cellular populations in this small dataset. Note that `ivis` is an “honest” algorithm and distances between observations are meaningful. Our benchmarks show that `ivis` is ~10% better at preserving local and global distances in low-dimensional space than comparable state-of-the-art algorithms. Additionally, `ivis` is robust against noise and outliers, ulike t-SNE, which tends to group random noise into well-defined clusters that can be potentially misleading.
