### 2019-07-31 beringresearch/ivis#38 @idroz
Hi! Really appreciate the feedback on this and the pull request. Will do the checks and update!
