### 2019-07-31 beringresearch/ivis#38 @idroz
Works great - merged #42 into master - thank you!
