### 2019-07-31 853aad0
add build vignette argument into install script
