### 2019-07-31 beringresearch/ivis#38 @kevinrue
With #43 I think that we can call this issue resolved.
