+++ ci_scripts/test.sh
@@ -1,0 +1,1 @@
+py.test tests
