### 2019-07-31 e058522
add travis-ci badge
