### 2019-08-01 7bccb9d
add instructions for virtualenv installation
