### 2019-08-01 8537e5d
add instructions for virtualenv installation
