### 2019-08-01 c633b54
Make syntax compatible with Python 2.
