### 2019-08-01 3328d96
add r package documentation to readthedocs
