### 2019-08-01 ff13e4b
add r package documentation to readthedocs
