### 2019-08-01 5cf5452
Add travis R scripts
