+++ ci_scripts/install-R.sh
@@ -1,0 +1,3 @@
+Rscript -e 'install.packages("reticulate")'
+Rscript -e 'install.packages("devtools")'
+Rscript -e 'devtools::install_github("beringresearch/ivis/R-package");library(ivis);install_ivis()'
