### 2019-08-01 c0bab5b
Add travis R scripts
