+++ ci_scripts/test-R.sh
@@ -1,0 +1,2 @@
+cd R-package
+Rscript -e 'library(devtools); test()'
