### 2019-08-01 7175a63
Fix Travis R installation
