+++ .travis.yml
@@ -13,7 +13,7 @@ matrix:
       os: linux
 
 before_install:
-  apt-get install -y R-core
+  apt-get install -y r-base r-base-dev
 
 install: 
   - source ci_scripts/install.sh
