### 2019-08-01 2586ac4
Add sudo to R install command
