### 2019-08-01 c46f45c
Execute Travis R package installation commands as sudo
