### 2019-08-01 42b6228
Add mirror to R package installation. Install python venv.
