### 2019-08-01 afa64da
update Rscripts string
