+++ ci_scripts/install-R.sh
@@ -1,5 +1,5 @@
 sudo Rscript -e 'install.packages("reticulate", repos="http://cran.us.r-project.org")'
-sudo Rscript -e 'install.packages("devtools"), repos="http://cran.us.r-project.org"'
+sudo Rscript -e 'install.packages("devtools", repos="http://cran.us.r-project.org")'
 
 python3 -m pip install --user virtualenv 
 
