### 2019-08-01 02faeb9
drop r version and user for travis
