+++ R-package/DESCRIPTION
@@ -8,7 +8,7 @@ Description: Implementation of the ivis algorithm as described in the paper:
   This algorithm uses a siamese neural network trained on triplets to reduce the dimensionality of data to two dimensions for visualization.
   Each triplet is sampled from one of the k nearest neighbours as approximated by the Annoy library, with neighbouring points being pulled
   together and non-neighours being pushed away.
-Depends: R (>= 3.4.3)
+Depends: R (>= 3.2.3)
 Imports: reticulate
 License: MIT 
 Encoding: UTF-8
