### 2019-08-01 0db3bda
drop r version and user for travis
