+++ ci_scripts/install-R.sh
@@ -1,6 +1,6 @@
 sudo Rscript -e 'install.packages("reticulate", repos="http://cran.us.r-project.org")'
 sudo Rscript -e 'install.packages("devtools", repos="http://cran.us.r-project.org")'
 
-python3 -m pip install --user virtualenv 
+python3 -m pip install virtualenv 
 
 sudo Rscript -e 'devtools::install_github("beringresearch/ivis/R-package");library(ivis);install_ivis()'
