+++ ci_scripts/install-R.sh
@@ -1,3 +1,4 @@
+deactivate
 sudo Rscript -e 'install.packages("reticulate", repos="http://cran.us.r-project.org")'
 sudo Rscript -e 'install.packages("devtools", repos="http://cran.us.r-project.org")'
 
