### 2019-08-01 974d622
Install latest version of reticulate
