+++ ci_scripts/install-R.sh
@@ -1,5 +1,5 @@
 deactivate
-sudo Rscript -e 'install.packages("reticulate", repos="http://cran.us.r-project.org")'
+sudo Rscript -e 'remotes::install_github("rstudio/reticulate")'
 sudo Rscript -e 'install.packages("devtools", repos="http://cran.us.r-project.org")'
 
 python3 -m pip install virtualenv 
