### 2019-08-01 f27d0ae
Add Numpy to Python3 Travis enviornment
