### 2019-08-01 0f8633a
Decactivate Travis default Python virutal environment before installing Python packages
