+++ ci_scripts/install.sh
@@ -1,2 +1,3 @@
+deactivate
 pip install -e .
 pip install pytest
