### 2019-08-01 6aa872d
Add Python 3 dependencies to new Travis Python3 non-virtual environment
