+++ ci_scripts/install.sh
@@ -1,3 +1,6 @@
 deactivate
+sudo apt-get install python3-pip
+python3 -m pip install --user --upgrade pip
+
 pip install -e .
 pip install pytest
