### 2019-08-01 0aae9bb
Add python-setuptools to Travis Python3 env
