### 2019-08-01 7a25c83
Add python-setuptools to Travis Python3 env
