### 2019-08-01 25b6141
Ensure Travis python package install uses pip3
