+++ ci_scripts/install.sh
@@ -1,2 +1,2 @@
-pip install -e .
-pip install pytest
+python3 -m pip install --user -e .
+python3 -m pip install --user pytest
