### 2019-08-02 beringresearch/ivis#33 @Szubie
We have set up Travis CI to build the latest version of the code and run tests for both the Python and R packages automatically. We've also added a badge in the README displays the results - currently, all tests are passing.
