### 2019-08-02 beringresearch/ivis#33 @kevinrue
Great work thanks!
With Travis CI in place, could you please add code coverage?
Sorry to bump the goal post. I am genuinely trying to help. You probably know how code coverage can save you time down the road in terms of development and debugging.

This is fairly straightforward at this point.
Allow me to use one of my repositories as an example (https://github.com/csoneson/iSEE)

I may forget a detail or two, but it should be as easy as:

Adding the following to your [.travis.yml](https://github.com/beringresearch/ivis/blob/master/.travis.yml)
```yaml
r_packages:
  - covr

after_success:
  - Rscript -e 'library(covr); codecov()'
```

Adding a `codecov.yml` file to your repository, here in an example:  https://github.com/csoneson/iSEE/blob/master/codecov.yml

To help you, you may want to use the CRAN [usethis](https://cran.r-project.org/web/packages/usethis/index.html) package, in particular the function `usethis::use_coverage()` to get template files rather than copy paste from the iSEE repository.

Thanks again for your responsiveness in this review process!
