+++ codecov.yaml
@@ -1,0 +1,20 @@
+coverage:
+    precision: 2
+    round: down
+    range: "70...100"
+
+    status:
+        project:
+            default:
+                target: auto
+                if_no_uploads: error
+
+    patch:
+        default:
+            if_no_uploads: error
+
+    changes: true
+
+comment:
+    layout: "header, diff, changes, suggestions"
+    behavior: default
