### 2019-08-02 0ac97be
modify R package path for coverage
