+++ .travis.yml
@@ -28,4 +28,4 @@ script:
   - bash ci_scripts/test.sh
   - bash ci_scripts/test-R.sh
 after_success:
-  - Rscript -e 'library(covr); codecov()'
+  - Rscript -e 'library(covr); codecov("R-package/")'
