+++ R-package/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(ivis)
+
+test_check("ivis")
