### 2019-08-02 beringresearch/ivis#33 @idroz
Thanks for the info! We added code coverage to [.travis.yml](https://github.com/beringresearch/ivis/blob/master/.travis.yml) and successfully tased sync with Codecov.io. Travis CI build is passing as well.
