### 2019-08-02 60d76b9
add patience to hyperparameter selection
