### 2019-08-02 beringresearch/ivis#33 @kevinrue
Great.
Here is the snippet to add the codecov badge to your README:

```
[![Codecov.io coverage status](https://codecov.io/github/beringresearch/ivis/coverage.svg?branch=master)](https://codecov.io/github/beringresearch/ivis)
```

[![Codecov.io coverage status](https://codecov.io/github/beringresearch/ivis/coverage.svg?branch=master)](https://codecov.io/github/beringresearch/ivis)

In the future, you can choose to either work on increase coverage where possible, or alternatively `covr` allows you to exclude certain parts of your code from the coverage testing. This should generally be reserved for parts of the code that _cannot_ be tested.
See https://cran.r-project.org/web/packages/covr/readme/README.html#exclusion-comments
