### 2019-08-02 beringresearch/ivis#33 @idroz
Nice! There were several things that we couldn't test - namely zzz.R and install_ivis.R. It's good to know that these can be excluded.

I did try to simply load the ``.onLoad()`` function as a means of testing zzz.R but since reticulate seems to enforce `<<-` operation on imported python objects, was unable to do so without raising an error.
