### 2019-08-03 19e719a
update hyperparameter selection
