+++ docs/hyperparameters.rst
@@ -17,64 +17,30 @@ Hyperparameter Selection
    created, followed by an embedding layer of size 'embedding\_dims'.
 
 ``k`` , ``n_epochs_without_progress``, and ``model`` are tunable parameters that should be selected on
-the basis of dataset size and complexity. We will look at each of these
-parameters in turn.
+the basis of dataset size and complexity.  The following table summarizes our findings:
 
+.. table:: Recommended hyperparameter settings for two-dimensional embeddings.
+    :widths: auto
 
-.. code:: ipython3
++----------------+--------+-------------------------------+------------+ 
+| Observations   |  ``k`` | ``n_epochs_without_progress`` | ``model``  |
++================+========+===============================+============+
+| < 1,000        | 10-15  | 20-30                         | "maaten"   |
++----------------+--------+-------------------------------+------------+
+| 1,000-10,000   | 10-15  | 10-20                         | "maaten"   |
++----------------+--------+-------------------------------+------------+
+| 10,000-50,000  | 15-150 | 10-20                         | "maaten"   |
++----------------+--------+-------------------------------+------------+
+| 50,000-100,000 | 15-150 | 10-15                         | "maaten"   |
++----------------+--------+-------------------------------+------------+
+| 100K - 500K    | 15-150 | 5-10                          | "maaten" |
++----------------+--------+-------------------------------+------------+
+| 500K - 1M      | 15-150 | 3-5                           | "default``|
++----------------+--------+-------------------------------+------------+
+| > 1M           | 15-150 | 2-3                           | +default``|
++----------------+--------+-------------------------------+------------+
 
-    import matplotlib.pyplot as plt
-    import mpl_toolkits.mplot3d.axes3d as p3
-    
-    from sklearn.datasets.samples_generator import make_swiss_roll
-    from sklearn.preprocessing import PolynomialFeatures, MinMaxScaler
-    
-    from sklearn.manifold import TSNE
-    from sklearn.decomposition import PCA
-    
-    import keras
-    import pydot as pyd
-    from IPython.display import SVG
-    from keras.utils.vis_utils import model_to_dot
-    
-    keras.utils.vis_utils.pydot = pyd
-    
-    from ivis import Ivis
-    from ivis.nn.network import base_network
-
-
-The Swiss Roll dataset with 10,000 points will be used to demonstrate
-the effects of ``k`` and ``model`` on embeddings.
-
-.. note::
-  Our aim is not to unfold the data in the Swiss Roll, but rather recover original structure from high-dimensional feature set.
-
-.. code:: ipython3
-
-    X, y = make_swiss_roll(n_samples=10000, noise=0.05)
-    X_poly = PolynomialFeatures(10).fit_transform(X)
-    X_poly = MinMaxScaler().fit_transform(X_poly)
-
-
-Since ``ivis`` is a dimensionality reduction algorithm, we will
-artificially add reduntant features to the original three-dimensional
-Swiss Roll dataset using polynomial combinations (degree ≤ 10) of the
-original features.
-   
-
-.. code:: ipython3
-
-    fig = plt.figure(figsize=(14, 10))
-    ax = fig.add_subplot(1, 1, 1, projection='3d')
-    
-    ax.view_init(7, -65)
-    ax.scatter(X[:, 0], X[:, 1], X[:, 2],
-                   c = y,
-                   s=1)
-    ax.set_title('Original')
-
-
-.. image:: _static/swiss_roll_original.png
+We will now look at each of these parameters in turn.
 
 
 ``k``
@@ -85,46 +51,21 @@ the dataset. Low ``k`` values will result in prioritisation of local
 dataset features and the overall global structure may be missed.
[…]
