### 2019-08-03 #1596 @kevinrue
@lpantano Both reviewers have ticked all our respective boxes. Let us know if you need anything else from us!
