### 2019-08-05 #1596 @lpantano
@openjournals/joss-eics, reviewers are done, please, let us know if this is ready to go. Thanks everybody!
