### 2019-08-05 #1596 @kevinrue
Argh, @lpantano was too fast for me. Here is what I was just typing:

As flattered as I am to be listed as a co-author on the citation in ZENODO
> Cite as:
> Ignat Drozdov, Benjamin Szubert, beringresearch, & Kevin Rue-Albrecht. (2019, August 5). beringresearch/ivis 1.2.3-joss (Version 1.2.3-joss). Zenodo. http://doi.org/10.5281/zenodo.3360569

I think what Lorena needs (if it's possible) would be to only have " Ignat Drozdov & Benjamin Szubert" as authors.
Also, "beringresearch" is an affiliation rather than an author.

With regards to my contributions to the R package during the review process, I think the fairest approach would be to add this to the R package DESCRIPTION file:
```
c(person("Kevin", "Rue-Albrecht", role = "ctb", email = "kevin.rue-albrecht@kennedy.ox.ac.uk", comment = c(ORCID = "0000-0003-3899-3872"))
```
It declares that I contributed to, rather than authored, the package.
See http://r-pkgs.had.co.nz/description.html#author
