### 2019-08-05 76c2fd2
add r contributors
