+++ R-package/DESCRIPTION
@@ -2,7 +2,8 @@ Package: ivis
 Title: Artificial neural network-driven visualization of high-dimensional data using triplets. 
 Version: 1.2.3
 Authors@R: c(person("Benjamin", "Szubert", email = "bszubert@beringresearch.com", role = c("aut", "cre")),
-         person("Ignat", "Drozdov", email = "idrozdov@beringresearch.com", role = c("aut")))
+         person("Ignat", "Drozdov", email = "idrozdov@beringresearch.com", role = c("aut")),
+         person("Kevin", "Rue-Albrecht", role = "ctb", email = "kevin.rue-albrecht@kennedy.ox.ac.uk", comment = c(ORCID = "0000-0003-3899-3872")))
 Description: Implementation of the ivis algorithm as described in the paper:
   'Structure-preserving visualisation of high dimensional single-cell data'.
   This algorithm uses a siamese neural network trained on triplets to reduce the dimensionality of data to two dimensions for visualization.
