### 2019-08-05 #1596 @idroz
Still getting to know my way around Zenodo! Author list should now be fixed. I have also modified R package DESCRIPTION file to acknowledge @kevinrue contribution. Thanks again!
