### 2019-08-05 #1596 @lpantano
Thank you @idroz, but you may need to bump the version and get another zenodo. Right now DESCRIPTION in master has @kevinrue  as author, but not the zenodo archive file. It would be possible to create a v 1.2.4 to make sure we point to the final version and the zenodo points to the final version? Thanks!
