### 2019-08-05 #1596 @idroz
v 1.2.4 released and author details synced across zenodo and GitHub.
