### 2019-08-05 #1596 @lpantano
Thanks, version looks good, and zenode as well, just need to update the title of the zenodo archive. The title should match the title of your paper (ivis: dimensionality reduction in very large datasets using Siamese Networks). thanks.
