### 2019-08-05 #1596 @lpantano
thanks!

ok, @openjournals/joss-eics , this time for real. I think all makes sense now. Sorry about the noise.
